Program Compatibility Conversion
When imported into Visual Studio Community 2015, the IDE asked to make a
"one-time upgrade" to the project. I presume that this was due to the
original program being developed in an earlier IDE version. With this
upgrade, the code is supposedly only compatible with my version of VS or
newer.